Работа над ошибками
