+++ lesson15-patterns1/src/main/java/ru/romansib/otus/Box.java
@@ -1,6 +1,8 @@
 package ru.romansib.otus;
 
+import java.util.Iterator;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public class Box {
     private final List<Integer> firstSection = List.of(1, 5, 9, 13, 17);
@@ -23,5 +25,82 @@ public class Box {
 
     public List<Integer> getFourthSection() {
         return fourthSection;
+    }
+
+    public void showBoxContent() {
+        BoxIterator iterator = new BoxIterator(this);
+        while (iterator.hasNext()) {
+            System.out.println(iterator.next());
+        }
+    }
+
+    private class BoxIterator implements Iterator<Integer> {
+        private int currentIndex;
+        private List<Integer> currentList;
+
+        private final List<Integer> firstSection;
+        private final List<Integer> secondSection;
+        private final List<Integer> thirdSection;
+        private final List<Integer> fourthSection;
+
+        public BoxIterator(Box box) {
+            this.firstSection = box.getFirstSection();
+            this.secondSection = box.getSecondSection();
+            this.thirdSection = box.getThirdSection();
+            this.fourthSection = box.getFourthSection();
+            this.currentIndex = getMaxIndex(box);
+            this.currentList = fourthSection;
+        }
+
+        private int getMaxIndex(Box box) {
+            int maxIndex = 0;
+            maxIndex = compare(box.getFirstSection(), maxIndex);
+            maxIndex = compare(box.getSecondSection(), maxIndex);
+            maxIndex = compare(box.getThirdSection(), maxIndex);
+            maxIndex = compare(box.getFourthSection(), maxIndex);
+            return maxIndex;
+        }
+
+        private int compare(List<Integer> l1, int maxIndex) {
+            if (l1.size() - 1 > maxIndex)
+                return l1.size() - 1;
+            return maxIndex;
+        }
+
+        @Override
+        public boolean hasNext() {
+            while (currentIndex >= 0) {
+                return true;
+            }
+            return false;
+        }
+
+        @Override
+        public Integer next() {
+            if (!hasNext()) {
+                throw new NoSuchElementException();
+            }
+            int result = currentList.get(currentIndex);
+            if (currentList == firstSection) {
+                currentIndex--;
+            }
+            switchList();
+            return result;
+        }
+
+        private void switchList() {
+            if (currentList.equals(firstSection)) {
+                currentList = fourthSection;
+            }
+            if (currentList.equals(secondSection)) {
+                currentList = firstSection;
+            }
+            if (currentList.equals(thirdSection)) {
+                currentList = secondSection;
+            }
+            if (currentList.equals(fourthSection)) {
+                currentList = thirdSection;
+            }
+        }
     }
 }
