+++ lesson15-patterns1/src/main/java/ru/romansib/otus/Product.java
@@ -2,18 +2,20 @@ package ru.romansib.otus;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class Product {
     private final long id;
     private final String title;
     private final String description;
-    private final Float cost;
+    private final BigDecimal cost;
     private final int weight;
     private final int width;
     private final int length;
     private final int height;
 
-    private Product(long id, String title, String description, Float cost, int weight, int width, int length, int height) {
+    private Product(long id, String title, String description, BigDecimal cost, int weight, int width, int length, int height) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -25,29 +27,11 @@ public class Product {
     }
 
 
-    public interface ProductBuilderI {
-        ProductBuilder id(int id);
-
-        ProductBuilder title(String title);
-
-        ProductBuilder description(String description);
-
-        ProductBuilder cost(Float cost);
-
-        ProductBuilder weight(int weight);
-
-        ProductBuilder width(int width);
-
-        ProductBuilder length(int length);
-
-        ProductBuilder height(int height);
-    }
-
-    public static final class ProductBuilder implements ProductBuilderI {
+    public static final class ProductBuilder {
         private long id;
         private String title;
         private String description;
-        private Float cost;
+        private BigDecimal cost;
         private int weight;
         private int width;
         private int length;
@@ -57,49 +41,42 @@ public class Product {
             return new Product(id, title, description, cost, weight, width, length, height);
         }
 
-        @Override
         public ProductBuilder id(int id) {
             this.id = id;
             return this;
         }
 
-        @Override
         public ProductBuilder title(String title) {
             this.title = title;
             return this;
         }
 
-        @Override
         public ProductBuilder description(String description) {
             this.description = description;
             return this;
         }
 
-        @Override
-        public ProductBuilder cost(Float cost) {
+
+        public ProductBuilder cost(BigDecimal cost) {
             this.cost = cost;
             return this;
         }
 
-        @Override
         public ProductBuilder weight(int weight) {
             this.weight = weight;
             return this;
         }
 
-        @Override
         public ProductBuilder width(int width) {
             this.width = width;
             return this;
         }
 
-        @Override
         public ProductBuilder length(int length) {
             this.length = length;
             return this;
         }
 
-        @Override
         public ProductBuilder height(int height) {
             this.height = height;
             return this;
